ai-event-planner/requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
google-generativeai==0.8.3
